* goo/GooString.cc:         * poppler/ABWOutputDev.cc: Build on Sun Force compiler, patch by         Darren Kenny <darren.kenny@sun.com>

+++ goo/GooString.cc
@@ -554,7 +554,7 @@ void GooString::formatDouble(double x, char *buf, int bufSize, int prec,
   if ((neg = x < 0)) {
     x = -x;
   }
-  x = floor(x * pow(10, prec) + 0.5);
+  x = floor(x * pow((double)10, prec) + 0.5);
   i = bufSize;
   started = !trim;
   for (j = 0; j < prec && i > 1; ++j) {
